move ServiceProvider and standardizeConfig into services-core
