+++ packages/services-core/src/index.js
@@ -1,3 +1,5 @@
+export { default as standardizeConfig } from './standardizeConfig';
+export { default as ServiceProvider } from './ServiceProvider';
 export { default as Container } from './Container';
 export { default as LocalService } from './LocalService';
 export { default as PublicService } from './PublicService';
